requirements.txt
arrow==1.2.3
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-datemath==1.5.5
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.2.3
